move AssetAliases to separate file
