+++ slk/chain/chain.py
@@ -6,7 +6,6 @@ from contextlib import contextmanager
 from pathlib import Path
 from typing import Any, Callable, Dict, Generator, List, Optional, Set, Union, cast
 
-from tabulate import tabulate
 from xrpl.models import (
     XRP,
     AccountInfo,
@@ -22,63 +21,11 @@ from xrpl.models import (
 from xrpl.models.transactions.transaction import Transaction
 from xrpl.utils import drops_to_xrp
 
+from slk.chain.asset_aliases import AssetAliases
 from slk.chain.key_manager import KeyManager
 from slk.chain.node import Node
 from slk.classes.common import Account
 from slk.classes.config_file import ConfigFile
-
-
-class AssetAliases:
-    def __init__(self: AssetAliases) -> None:
-        self._aliases: Dict[str, IssuedCurrency] = {}  # alias -> IssuedCurrency
-
-    def add(self: AssetAliases, asset: IssuedCurrency, name: str) -> None:
-        self._aliases[name] = asset
-
-    def is_alias(self: AssetAliases, name: str) -> bool:
-        return name in self._aliases
-
-    def asset_from_alias(self: AssetAliases, name: str) -> IssuedCurrency:
-        assert name in self._aliases
-        return self._aliases[name]
-
-    def known_aliases(self: AssetAliases) -> List[str]:
-        return list(self._aliases.keys())
-
-    def known_assets(self: AssetAliases) -> List[IssuedCurrency]:
-        return list(self._aliases.values())
-
-    def to_string(self: AssetAliases, nickname: Optional[str] = None) -> str:
-        data = []
-        if nickname:
-            if nickname in self._aliases:
-                v = self._aliases[nickname]
-                currency = v.currency
-                issuer = v.issuer if v.issuer else ""
-            else:
-                currency = "NA"
-                issuer = "NA"
-            data.append(
-                {
-                    "name": nickname,
-                    "currency": currency,
-                    "issuer": issuer,
-                }
-            )
-        else:
-            for (k, v) in self._aliases.items():
-                data.append(
-                    {
-                        "name": k,
-                        "currency": v.currency,
-                        "issuer": v.issuer if v.issuer else "",
-                    }
-                )
-        return tabulate(
-            data,
-            headers="keys",
-            tablefmt="presto",
-        )
 
 
 class Chain:
